+++ app/src/main/java/com/neqabty/domain/entities/MedicalRenewalPaymentEntity.kt
@@ -8,7 +8,7 @@ data class MedicalRenewalPaymentEntity(
 ) {
     data class PaymentItem(
             var paymentRequestNumber: String? = "",
-            var amount: Int?,
+            var amount: Double?,
             var name: String?,
             var engName: String?,
             var engNumber: String?,
